wdpress not working

+++ srcs/requirements/wordpress/tools/setup-wdpress.sh
@@ -1,35 +1,55 @@
-sed -i 's/listen = 127.0.0.1:9000/listen = 9000/g' /etc/php/7.4/fpm/pool.d/www.conf
-sed -i 's/;listen.owner = /listen.owner = www-data/g' /etc/php/7.4/fpm/pool.d/www.conf
-sed -i 's/;listen.group = nobody/listen.group = www-data/g' /etc/php/7.4/fpm/pool.d/www.conf
+cat > etc/php/7.4/fpm/pool.d/www.conf << EOF
+[www]
+user = www-data
+group = www-data
+listen = wordpress:9000
+pm = dynamic
+pm.start_servers = 6
+pm.max_children = 25
+pm.min_spare_servers = 2
+pm.max_spare_servers = 10
+EOF
 
 mkdir -p /var/www
-chmod -R 777 /var/www
 
-wget https://wordpress.org/latest.zip
-unzip latest.zip -d /var/www/
-rm latest.zip
-mv /var/www/wordpress/* /var/www/
-rm -rf /var/www/wordpress
+#wget https://wordpress.org/latest.zip
+#unzip latest.zip
+#rm latest.zip
+#mv /wordpress/* /var/www/
+#rm -rf wordpress
 
-mkdir ~/data
-mkdir ~/data/mariadb
-mkdir ~/data/wordpress
-chmod -R 777 ~/data
+#sed -i "s/listen = \/run\/php\/php7.4-fpm.sock/listen = 9000/" "/etc/php/7.4/fpm/pool.d/www.conf";
 
-#touch wp-config.php
-echo "<?php" > /var/www/wp-config.php
-echo "define( 'DB_NAME', '${DB_NAME}' );" >> /var/www/wp-config.php
-echo "define( 'DB_USER', '${DB_USER}' );" >> /var/www/wp-config.php
-echo "define( 'DB_PASSWORD', '${DB_PASS}' );" >> /var/www/wp-config.php
-echo "define( 'DB_HOST', 'mariadb' );" >> /var/www/wp-config.php
-echo "define( 'DB_CHARSET', 'utf8' );" >> /var/www/wp-config.php
-echo "define( 'DB_COLLATE', '' );" >> /var/www/wp-config.php
-echo "define('FS_METHOD','direct');" >> /var/www/wp-config.php
-echo "\$table_prefix = 'wp_';" >> /var/www/wp-config.php
-echo "define( 'WP_DEBUG', false );" >> /var/www/wp-config.php
-echo "if ( ! defined( 'ABSPATH' ) ) {" >> /var/www/wp-config.php
-echo "define( 'ABSPATH', __DIR__ . '/' );}" >> /var/www/wp-config.php
-echo "require_once ABSPATH . 'wp-settings.php';" >> /var/www/wp-config.php
-echo "?>" >> /var/www/wp-config.php
+mkdir -p /run/php/;
+touch /run/php/php7.4-fpm.pid;
 
-#chown -R www-data:www-data /var/www
+if [ ! -f /var/www/wp-config.php ]; then
+	cd /var/www/
+	touch /var/www/wp-config.php
+	echo "<?php" > /var/www/wp-config.php
+	echo "define( 'DB_NAME', '${DB_NAME}' );" >> /var/www/wp-config.php
+	echo "define( 'DB_USER', '${DB_USER}' );" >> /var/www/wp-config.php
+	echo "define( 'DB_PASSWORD', '${DB_PASS}' );" >> /var/www/wp-config.php
+	echo "define( 'DB_HOST', '${DB_HOST}' );" >> /var/www/wp-config.php
+	echo "define( 'DB_CHARSET', 'utf8' );" >> /var/www/wp-config.php
+	echo "define( 'DB_COLLATE', '' );" >> /var/www/wp-config.php
+	curl https://api.wordpress.org/secret-key/1.1/salt/ > \
+	echo >> /var/www/wp-config.php
+	echo "define('FS_METHOD','direct');" >> /var/www/wp-config.php
+	echo "\$table_prefix = 'wp_';" >> /var/www/wp-config.php
+	echo "define( 'WP_DEBUG', true );" >> /var/www/wp-config.php
+	echo "if ( ! defined( 'ABSPATH' ) ) {" >> /var/www/wp-config.php
+	echo "define( 'ABSPATH', __DIR__ . '/' );}" >> /var/www/wp-config.php
+	echo "require_once ABSPATH . 'wp-settings.php';" >> /var/www/wp-config.php
+	#echo "?>" >> /var/www/wp-config.php
+	echo "Wordpress: setting up..."
+	wget https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar;
+	chmod +x wp-cli.phar; 
+	mv wp-cli.phar /usr/local/sbin/wp;
+	wp core download --allow-root;
+	echo "Wordpress: creating users..."
+	wp core install --allow-root --url=$SERVER_NAME --title=$DB_NAME --admin_user=${DB_USER} --admin_password=${DB_PASS} --admin_email=mail@mail.com
+	#wp user create --allow-root ${DB_ROOT_USER} mail@mail.com --user_pass=${DB_ROOT_PASS};
+	echo "Wordpress: set up!"
+fi
+
